+++ spec/ultra_settings/configuration_view_spec.rb
@@ -22,10 +22,8 @@ describe UltraSettings::ConfigurationView do
 
   it "renders valid HTML", env: {TEST_STRING: "<script"} do
     html = UltraSettings::ConfigurationView.new(TestConfiguration.instance).render
-    docs = parse_with_svg(html)
-    docs.each do |doc|
-      expect(doc.errors).to be_empty
-    end
+    doc = Nokogiri::HTML5(html)
+    expect(doc.errors).to be_empty
   end
 
   describe "links to runtime settings" do