+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Show icon with a tooltip with the current value for each data source in the web UI.
+- Show icons on the web UI that open a dialog with the current value for each data source.
 
 ## 2.6.0